let 객체 삭제 및 return 수정 변경
